api/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
aiohttp==3.8.1
firebase-admin==5.1.0
pyTelegramBotAPI==4.7.0
requests==2.26.0
python-dotenv==1.0.0
